end of section 28

+++ codehacking/app/Http/Controllers/AdminUsersController.php
@@ -143,7 +143,10 @@ class AdminUsersController extends Controller
 
         $user = User::findOrFail($id);
 
-        unlink(public_path(). $user->photo->file);
+        if($user->photo){
+
+            unlink(public_path(). $user->photo->file);
+        }
         $user->delete();
 
         Session::flash('deleted_user', 'The user has been deleted');
